+++ reading-list-app/reading-list-front-end-with-managed-auth/src/api/books/delete-books.ts
@@ -16,7 +16,7 @@
 
 
 import { apiUrl } from "./constants";
-import { performRequest } from "../retry";
+import { performRequest } from "../request-wrapper";
 
 
 export async function deleteBooks(id: string) {
